+++ traffic_ops/traffic_ops_golang/api/api.go
@@ -41,6 +41,7 @@ import (
 	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/config"
 	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/tocookie"
 
+	influx "github.com/influxdata/influxdb1-client/v2"
 	"github.com/jmoiron/sqlx"
 	"github.com/lib/pq"
 )
@@ -49,6 +50,18 @@ const DBContextKey = "db"
 const ConfigContextKey = "context"
 const ReqIDContextKey = "reqid"
 const APIRespWrittenKey = "respwritten"
+
+const influxServersQuery = `
+SELECT (host_name||'.'||domain_name) as fqdn,
+       tcp_port,
+       https_port
+FROM server
+WHERE type in ( SELECT id
+                FROM type
+                WHERE name='INFLUXDB'
+              )
+AND status=(SELECT id FROM status WHERE name='ONLINE')
+`
 
 // WriteResp takes any object, serializes it as JSON, and writes that to w. Any errors are logged and written to w via tc.GetHandleErrorsFunc.
 // This is a helper for the common case; not using this in unusual cases is perfectly acceptable.
@@ -298,7 +311,7 @@ type APIInfo struct {
 func DeprecationWarning(alternative string) tc.Alert {
 	return tc.Alert{
 		Level: tc.WarnLevel.String(),
-		Text: fmt.Sprintf("This request method of this endpoint is deprecated. You are advised to switch to '%s' at your earliest convenience", alternative),
+		Text:  fmt.Sprintf("This request method of this endpoint is deprecated. You are advised to switch to '%s' at your earliest convenience", alternative),
 	}
 }
 
@@ -404,6 +417,66 @@ func SendMail(to rfc.EmailAddress, msg []byte, cfg *config.Config) (int, error, 
 		return http.StatusInternalServerError, nil, fmt.Errorf("Failed to send email: %v", err)
 	}
 	return http.StatusOK, nil, nil
+}
+
+// CreateInfluxClient onstructs and returns an InfluxDB HTTP client, if enabled and when possible.
+// The error this returns should not be exposed to the user; it's for logging purposes only.
+//
+// If Influx connections are not enabled, this will return `nil` - but also no error. It is expected
+// that the caller will handle this situation appropriately.
+func (inf *APIInfo) CreateInfluxClient() (*influx.Client, error) {
+	if !inf.Config.InfluxEnabled {
+		return nil, nil
+	}
+
+	var fqdn string
+	var tcpPort uint
+	var httpsPort sql.NullInt64 // this is the only one that's optional
+
+	row := inf.Tx.Tx.QueryRow(influxServersQuery)
+	if e := row.Scan(&fqdn, &tcpPort, &httpsPort); e != nil {
+		return nil, fmt.Errorf("Failed to create influx client: %v", e)
+	}
+
+	host := "http%s://%s:%d"
+	if inf.Config.ConfigInflux != nil && *inf.Config.ConfigInflux.Secure {
+		if !httpsPort.Valid {
+			log.Warnf("INFLUXDB Server %s has no secure ports, assuming default of 8086!", fqdn)
+			httpsPort = sql.NullInt64{8086, true}
+		}
+		port, err := httpsPort.Value()
+		if err != nil {
+			return nil, fmt.Errorf("Failed to create influx client: %v", err)
+		}
+
+		p := port.(int64)
+		if p <= 0 || p > 65535 {
+			log.Warnf("INFLUXDB Server %s has invalid port, assuming default of 8086!", fqdn)
+			p = 8086
+		}
+
+		host = fmt.Sprintf(host, "s", fqdn, p)
+	} else if tcpPort > 0 && tcpPort <= 65535 {
+		host = fmt.Sprintf(host, "", fqdn, tcpPort)
+	} else {
+		log.Warnf("INFLUXDB Server %s has invalid port, assuming default of 8086!", fqdn)
+		host = fmt.Sprintf(host, "", fqdn, 8086)
+	}
+
+	config := influx.HTTPConfig{
+		Addr:      host,
+		Username:  inf.Config.ConfigInflux.User,
+		Password:  inf.Config.ConfigInflux.Password,
+		UserAgent: fmt.Sprintf("TrafficOps/%s (Go)", inf.Config.Version),
+		Timeout:   time.Duration(float64(inf.Config.ReadTimeout)/2.1) * time.Second,
+	}
+
+	var client influx.Client
+	client, e := influx.NewHTTPClient(config)
+	if client == nil {
+		return nil, fmt.Errorf("Failed to create influx client (client was nil): %v", e)
+	}
+	return &client, e
 }
 
 // APIInfoImpl implements APIInfo via the APIInfoer interface